fix: address PR #46 high-priority review findings (H1 + H2)
H1: Route handle_google_callback to server root, bypassing /api prefix.
H2: Extract _SAFE_ID_RE + _validate_safe_id to shared _validation.py,
    update token.py and proxy.py to import from it.
